requirements.txt
APScheduler==3.7.0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
Flask==2.0.1
Flask-APScheduler==1.12.2
Flask-SQLAlchemy==2.4.4
google-auth==1.30.1
greenlet==1.1.0
gunicorn==20.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.3.23
tzlocal==2.1
urllib3==1.26.5
Werkzeug==2.0.1
